organize analysis files
Main and sensitivity analyses are broken up into their own files now and
all call a common setup file. In the setup file, objects are created that contain both the
naive prediction risks on the original scale and on the natural log
scale. These are for use in the various risk plots.